UPSTREAM: x86/apic: Remove yet another dubious callback
Paranoia is not wrong, but having an APIC callback which is in most
implementations a complete NOOP and in one actually looking whether the
APICID of an upcoming CPU has been registered. The same APICID which was
used to bring the CPU out of wait for startup.

That's paranoia for the paranoia sake. Remove the voodoo.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Tested-by: Michael Kelley <mhklinux@outlook.com>
Tested-by: Sohil Mehta <sohil.mehta@intel.com>
Link: https://lore.kernel.org/r/20240212154640.116510935@linutronix.de
(cherry picked from commit 4a5f72a4a39f5d5dcf9b9dc1acc57ecbbb8d4caa)

BUG=b:349369441
TEST='cat /sys/devices/system/cpu/smt/control' on nautilus yields
     'notsupported'

Signed-off-by: Paweł Anikiel <panikiel@google.com>