UPSTREAM: dt-bindings: mfd: cros-ec: Add properties for GPIO controller
The ChromeOS embedded controller (EC) supports setting the state of
GPIOs when the system is unlocked, and getting the state of GPIOs in all
cases. The GPIOs are on the EC itself, so the EC acts similar to a GPIO
expander. Add the #gpio-cells and gpio-controller properties to the
cros-ec binding so that other devices described in DT can get the GPIOs
on the EC.

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/20240219202325.4095816-2-swboyd@chromium.org
Signed-off-by: Lee Jones <lee@kernel.org>
(cherry picked from commit 7b79740d42e7535901296f7acbab1043633422e6)

BUG=b:326655441
TEST=gpioinfo on Trogdor shows EC gpio device

Change-Id: I1d56108cb4a934a4dad0913704eaa5b4d2a62e52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5437875
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>