UPSTREAM: gpio: Add ChromeOS EC GPIO driver
The ChromeOS embedded controller (EC) supports setting the state of
GPIOs when the system is unlocked, and getting the state of GPIOs in all
cases. The GPIOs are on the EC itself, so the EC acts similar to a GPIO
expander. Add a driver to get and set the GPIOs on the EC through the
host command interface.

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>
(cherry picked from commit f837fe1bffe6976ed5e7198b1892ea248b75358b)

BUG=b:326655441
TEST=gpioinfo on Trogdor shows EC gpio device

Change-Id: Ia8116ffdf482ededd0f993fa5a8ac55c6c3bc17a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5437667
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>