FROMLIST: clk: Don't hold prepare_lock when calling kref_put()
We don't need to hold the prepare_lock when dropping a ref on a struct
clk_core. The release function is only freeing memory and any code with
a pointer reference has already unlinked anything pointing to the
clk_core. This reduces the holding area of the prepare_lock a bit.

Note that we also don't call free_clk() with the prepare_lock held.
There isn't any reason to do that.

Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(am from https://lore.kernel.org/r/20240325184204.745706-3-sboyd@kernel.org)

BUG=b:328070191, b:330474631
TEST=Boot on lazor
     Recovery flow on lazor
UPSTREAM-TASK=b:331232181

Change-Id: I5d2fcf7f9c5ad81426cc9201e3f4200688abbb09
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5389707
Reviewed-by: Douglas Anderson <dianders@chromium.org>