BACKPORT: FROMGIT: clk: Get runtime PM before walking tree for clk_summary
Similar to the previous commit, we should make sure that all devices are
runtime resumed before printing the clk_summary through debugfs. Failure
to do so would result in a deadlock if the thread is resuming a device
to print clk state and that device is also runtime resuming in another
thread, e.g the screen is turning on and the display driver is starting
up. We remove the calls to clk_pm_runtime_{get,put}() in this path
because they're superfluous now that we know the devices are runtime
resumed. This also squashes a bug where the return value of
clk_pm_runtime_get() wasn't checked, leading to an RPM count underflow
on error paths.

Fixes: 1bb294a7981c ("clk: Enable/Disable runtime PM for clk_summary")
Cc: Taniya Das <quic_tdas@quicinc.com>
Cc: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
Link: https://lore.kernel.org/r/20240325184204.745706-6-sboyd@kernel.org
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 9d1e795f754db1ac3344528b7af0b17b8146f321
 https://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git clk-fixes)

Conflicts:
   drivers/clk/clk.c

We don't have commit dcce5cc7826e ("clk: Show active consumers of clocks
in debugfs") but the conflict is a printk that is unaffected by this
patch.

BUG=b:331232181
TEST=git diff

Change-Id: I96240996562ba42e047675fcbd0f1d06ac5d2a18
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5449580
Reviewed-by: Douglas Anderson <dianders@chromium.org>