BACKPORT: FROMLIST: clk: Get runtime PM before walking tree for clk_summary
Similar to the previous commit, we should make sure that all devices are
runtime resumed before printing the clk_summary through debugfs. Failure
to do so would result in a deadlock if the thread is resuming a device
to print clk state and that device is also runtime resuming in another
thread, e.g the screen is turning on and the display driver is starting
up. We remove the calls to clk_pm_runtime_{get,put}() in this path
because they're superfluous now that we know the devices are runtime
resumed. This also squashes a bug where the return value of
clk_pm_runtime_get() wasn't checked, leading to an RPM count underflow
on error paths.

Fixes: 1bb294a7981c ("clk: Enable/Disable runtime PM for clk_summary")
Cc: Taniya Das <quic_tdas@quicinc.com>
Cc: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(am from https://patchwork.kernel.org/patch/13602755/)
(also found at https://lore.kernel.org/r/20240325184204.745706-6-sboyd@kernel.org)

Conflicts:
   drivers/clk/clk.c

We don't have commit dcce5cc7826e ("clk: Show active consumers of clocks
in debugfs") but the conflict is a printk that is unaffected by this
patch.

BUG=b:328070191, b:330474631
TEST=Boot on lazor
     Recovery flow on lazor
UPSTREAM-TASK=b:331232181

Change-Id: Icf8a57baf2f3fd697de81c70a5dc84548781d8b1
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5393465
Reviewed-by: Douglas Anderson <dianders@chromium.org>