UPSTREAM: crypto: x86/aes-xts - wire up VAES + AVX2 implementation
Add an AES-XTS implementation "xts-aes-vaes-avx2" for x86_64 CPUs with
the VAES, VPCLMULQDQ, and AVX2 extensions, but not AVX512 or AVX10.
This implementation uses ymm registers to operate on two AES blocks at a
time.  The assembly code is instantiated using a macro so that most of
the source code is shared with other implementations.

This is the optimal implementation on AMD Zen 3.  It should also be the
optimal implementation on Intel Alder Lake, which similarly supports
VAES but not AVX512.  Comparing to xts-aes-aesni-avx on Zen 3,
xts-aes-vaes-avx2 provides 70% higher AES-256-XTS decryption throughput
with 4096-byte messages, or 23% higher with 512-byte messages.

A large improvement is also seen with CPUs that do support AVX512 (e.g.,
98% higher AES-256-XTS decryption throughput on Ice Lake with 4096-byte
messages), though the following patches add AVX512 optimized
implementations to get a bit more performance on those CPUs.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit e787060bdfa35f8b40ef4d277a345ee35b41039f)
BUG=b:349650416
TEST=Applied full stack of CLs and ran crypto self-tests on Ice Lake
Change-Id: I50e54d01e26c37ef9d400d4db22dd55ec5d10e63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5661505
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>