UPSTREAM: crypto: x86/aes-xts - handle CTS encryption more efficiently
When encrypting a message whose length isn't a multiple of 16 bytes,
encrypt the last full block in the main loop.  This works because only
decryption uses the last two tweaks in reverse order, not encryption.

This improves the performance of decrypting messages whose length isn't
a multiple of the AES block length, shrinks the size of
aes-xts-avx-x86_64.o by 5.0%, and eliminates two instructions (a test
and a not-taken conditional jump) when encrypting a message whose length
*is* a multiple of the AES block length.

While it's not super useful to optimize for ciphertext stealing given
that it's rarely needed in practice, the other two benefits mentioned
above make this optimization worthwhile.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 1d27e1f5c8f7ade40f0e85ddecbe9158393265e5)
BUG=b:349650416
TEST=Applied full stack of CLs and ran crypto self-tests on Ice Lake
Change-Id: Id38951ebda2e2587bff20fba29b633599174d1a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5661510
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>