UPSTREAM: crypto: x86/aes-xts - handle AES-128 and AES-192 more efficiently
Decrease the amount of code specific to the different AES variants by
"right-aligning" the sequence of round keys, and for AES-128 and AES-192
just skipping irrelevant rounds at the beginning.

This shrinks the size of aes-xts-avx-x86_64.o by 13.3%, and it improves
the efficiency of AES-128 and AES-192.  The tradeoff is that for AES-256
some additional not-taken conditional jumps are now executed.  But these
are predicted well and are cheap on x86.

Note that the ARMv8 CE based AES-XTS implementation uses a similar
strategy to handle the different AES variants.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 2717e01fc3fb4d37b625b9bd6cf161d0d9d5c4b5)
BUG=b:349650416
TEST=Applied full stack of CLs and ran crypto self-tests on Ice Lake
Change-Id: I89662b6128ba72cfef9e900b508bafe33e52d4c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5661512
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>